+++ DataFormats/GeometryVector/src/TrackingJacobians.cc
@@ -1,4 +1,4 @@
-#include "DataFormats/GeometryVector/interface/jacobians.h"
+#include "DataFormats/GeometryVector/interface/TrackingJacobians.h"
 
 // Code moved from TrackingTools/AnalyticalJacobians
