+++ TrackingTools/AnalyticalJacobians/src/JacobianCartesianToCurvilinear.cc
@@ -1,6 +1,6 @@
 #include "TrackingTools/AnalyticalJacobians/interface/JacobianCartesianToCurvilinear.h"
 #include "TrackingTools/TrajectoryParametrization/interface/GlobalTrajectoryParameters.h"
-#include "DataFormats/GeometryVector/interface/jacobians.h"
+#include "DataFormats/GeometryVector/interface/TrackingJacobians.h"
 
 JacobianCartesianToCurvilinear::
 JacobianCartesianToCurvilinear(const GlobalTrajectoryParameters& globalParameters) :
